test(db-conformance): ported audit, counters and admin-store suites
